+++ st2common/st2common/exceptions/content.py
@@ -20,5 +20,5 @@ class UnsupportedMetaException(StackStormBaseException):
     pass
 
 
-class ParseException(StackStormBaseException):
+class ParseException(ValueError):
     pass