+++ st2common/st2common/models/utils/action_alias_utils.py
@@ -14,10 +14,14 @@
 # limitations under the License.
 
 import re
-from st2common.exceptions import content
+
+from st2common.exceptions.content import ParseException
 
 __all__ = [
-    'ActionAliasFormatParser'
+    'ActionAliasFormatParser',
+
+    'extract_parameters_for_action_alias_db',
+    'extract_parameters',
 ]
 
 
@@ -28,8 +32,14 @@ class ActionAliasFormatParser(object):
         self._param_stream = param_stream or ''
 
     def get_extracted_param_value(self):
+        """
+        Match command against the format string and extract paramters from the command string.
 
+        :rtype: ``dict``
+        """
         result = {}
+
+        param_stream = self._param_stream
 
         # As there's a lot of questions about using regular expressions,
         # I'll try to be thorough when documenting this code.
@@ -64,11 +74,12 @@ class ActionAliasFormatParser(object):
         # 1. Matching the arbitrary key-value pairs at the end of the command
         # to support extra parameters (not specified in the format string),
         # and cutting them from the command string afterwards.
-        ending_pairs = re.match(snippets['ending'], self._param_stream, re.DOTALL)
-        if ending_pairs:
+        ending_pairs = re.match(snippets['ending'], param_stream, re.DOTALL)
+        has_ending_pairs = ending_pairs and ending_pairs.group(1)
+        if has_ending_pairs:
             kv_pairs = re.findall(snippets['pairs'], ending_pairs.group(1), re.DOTALL)
-            self._param_stream = self._param_stream.replace(ending_pairs.group(1), '')
-        self._param_stream = " %s " % self._param_stream
+            param_stream = param_stream.replace(ending_pairs.group(1), '')
+        param_stream = " %s " % (param_stream)
 
         # 2. Matching optional parameters (with default values).
         optional = re.findall(snippets['optional'], self._format, re.DOTALL)
@@ -82,21 +93,56 @@ class ActionAliasFormatParser(object):
         reg = '^\s*' + reg + r'\s*$'
 
         # 3. Matching the command against our regex to get the param values
-        matched_stream = re.match(reg, self._param_stream, re.DOTALL)
+        matched_stream = re.match(reg, param_stream, re.DOTALL)
+
+        if not matched_stream:
+            # If no match is found we throw since this indicates provided user string (command)
+            # didn't match the provided format string
+            raise ParseException('Command "%s" doesn\'t match format string "%s"' %
+                                 (self._param_stream, self._format))
 
         # Compiling results from the steps 1-3.
         if matched_stream:
             result = matched_stream.groupdict()
+
         for param in optional:
             matched_value = result[param[0]] if matched_stream else None
             matched_result = matched_value or ''.join(param[1:])
             if matched_result is not None:
                 result[param[0]] = matched_result
-        if ending_pairs:
+
+        if has_ending_pairs:
             for pair in kv_pairs:
                 result[pair[0]] = ''.join(pair[2:])
 
         if self._format and not (self._param_stream.strip() or any(result.values())):
-            raise content.ParseException('No value supplied and no default value found.')
+            raise ParseException('No value supplied and no default value found.')
 
         return result
+
+
+def extract_parameters_for_action_alias_db(action_alias_db, format_str, param_stream):
+    """
+    Extract parameters from the user input based on the provided format string.
+
+    Note: This function makes sure that the provided format string is indeed available in the
+    action_alias_db.formats.
+    """
+    formats = []
+    for formatstring in action_alias_db.formats:
+        if isinstance(formatstring, dict) and formatstring.get('representation'):
+            formats.extend(formatstring['representation'])
+        else:
+            formats.append(formatstring)
+
+    if format_str not in formats:
+        raise ValueError('Format string "%s" is not available on the alias "%s"' %
+                         (format_str, action_alias_db.name))
+
+    result = extract_parameters(format_str=format_str, param_stream=param_stream)
+    return result
+
+
+def extract_parameters(format_str, param_stream):
+    parser = ActionAliasFormatParser(alias_format=format_str, param_stream=param_stream)
+    return parser.get_extracted_param_value()